+++ index.html
@@ -113,8 +113,8 @@
 
     <div id="total" class="activities blue block hidden">
         <div id="uv-div" class="component"></div>
-        <h1 class="component text-right scale-transition scale-out" id="scrollfire1">You have travelled <u id="totalMiles" class="red-text">0</u> since <u id="firstDate" class="red-text">0</u></h1>
-        <h2 class="component text-right scale-transition scale-out" style="top:50%" id="scrollfire2">That is an average of <u id="averageMiles" class="red-text">0</u> <br> miles a day</h1>
+        <h1 class="component text-right scale-transition scale-out" id="scrollfire1">You have travelled <u id="totalKm" class="red-text">0</u> since <u id="firstDate" class="red-text">0</u></h1>
+        <h2 class="component text-right scale-transition scale-out" style="top:50%" id="scrollfire2">That is an average of <u id="averageKm" class="red-text">0</u> <br> km a day</h1>
   	<div id="clouds">
 		<div class="cloud x1"></div>
 		<!-- Time for multiple clouds to dance around -->
@@ -126,12 +126,12 @@
   </div>
   <div id="flying" class="activities blue block hidden">
   	<div id="bottomCloud" class="component"></div>
-  	<h1 class="component text-right scale-transition scale-out" id="scrollfire3">You have flown <u id="planeTrips" class="red-text">0</u> times which has amounted to <u id="planeMiles" class="red-text">0</u> miles</h1>
+  	<h1 class="component text-right scale-transition scale-out" id="scrollfire3">You have flown <u id="planeTrips" class="red-text">0</u> times which has amounted to <u id="planeKm" class="red-text">0</u> km</h1>
   </div>
   <div id="driven" class="activities green block hidden">
   	<div id="road" class="component"></div>
   	<div id="road-sign" class="component">
-  		<h5 class="white-text" style="margin-left: 1rem; margin-right: 210px">TOTAL MILES DRIVEN <hr> <p id="milesDriven">0</p></h5>
+  		<h5 class="white-text" style="margin-left: 1rem; margin-right: 210px">TOTAL KM DRIVEN <hr> <p id="kmDriven">0</p></h5>
   	</div>
   	<h1 class="component text-left scale-transition scale-out" id="scrollfire4">You have gotten into a car <u id="inCar" class="red-text">0</u> times</h1>
     <h1 class="component text-right scale-transition scale-out" id="scrollfire5" style="top:50%">That is <u id="drivenEarth" class="red-text">0</u> times around <u class="red-text">the earth</u></h1>
@@ -139,9 +139,9 @@
   </div>
   <div id="onFoot" class="activities brown lighten-1 block hidden">
   	<div id="path" class="component"></div>
-  	<h1 class="component text-left scale-transition scale-out" id="scrollfire6" style="top:25%">You have walked <u id="walked" class="red-text">0</u> miles</h1>
-  	<h1 class="component text-right scale-transition scale-out" id="scrollfire7" style="top:45%">You have run <u id="ran" class="red-text">0</u> miles</h1>
-  	<h1 class="component text-left scale-transition scale-out" id="scrollfire8" style="top:65%">You have biked <u id="biked" class="red-text">0</u> miles</h1>
+  	<h1 class="component text-left scale-transition scale-out" id="scrollfire6" style="top:25%">You have walked <u id="walked" class="red-text">0</u> km</h1>
+  	<h1 class="component text-right scale-transition scale-out" id="scrollfire7" style="top:45%">You have run <u id="ran" class="red-text">0</u> km</h1>
+  	<h1 class="component text-left scale-transition scale-out" id="scrollfire8" style="top:65%">You have biked <u id="biked" class="red-text">0</u> km</h1>
   </div>
 
     </body>
